*model state can now be saved also when parallel execution is enabled
git-svn-id: http://svn.geogr.uni-jena.de/svn/jams/trunk@2957 cee23a32-52fb-0310-bcb1-fe94ad5b3aa2

+++ JAMSmain/src/jams/workspace/stores/DefaultOutputDataStore.java
@@ -178,7 +178,7 @@ public class DefaultOutputDataStore implements OutputDataStore {
         if (writer != null)
             out.writeLong(writer.getPosition());
         else
-            out.writeLong(0L);
+            out.writeLong((long)0L);
     }
 
     @Override
